requirements.txt
climate-assessment==0.1.3
pooch==1.8.0
pyam-iamc==1.7.0
